requirements.txt
psycopg2-binary==2.9.10
redis==5.0.8
python-dotenv==1.0.1
tabulate==0.9.0
